Add a prop to prevent scaling the canvas based on the device pixel ratio
The default behavior is to resize and scale the canvas up based on the
device pixel ratio (using `window.devicePixelRatio`). For example, if
the ratio is 3, the `width` and `height` of the canvas will be
multiplied by 3. This causes the output to be 3x the size as well,
which can be undesired.

This change adds a `scaleToDevicePixelRatio` prop, which defaults to
`true`, but allows setting it to `false` which will prevent scaling
the canvas.